+++ src/lib/components/Separator.svelte
@@ -1,1 +1,1 @@
-<hr class="bg-ui-700 shrink-0 h-[2px] w-full" />
+<hr />
