+++ src/lib/components/charts/doughnut-chart.svelte
@@ -9,7 +9,7 @@
 		type ChartConfiguration
 	} from 'chart.js';
 	import colors from 'tailwindcss/colors';
-	import { TOOLTIP } from '$lib/constants/chartConfig';
+	import { TOOLTIP } from '$constants/chart-config';
 	import { COLOURS } from '$constants/colours';
 
 	interface Props {
