+++ src/lib/components/charts/legend-list.svelte
