+++ src/lib/components/inputs/currency-input.svelte
