+++ src/lib/components/inputs/percentage-input.svelte
