--- src/lib/components/layout/Footer.svelte
@@ -1,9 +1,0 @@
-<script lang="ts">
-	let year = new Date().getFullYear();
-</script>
-
-<footer class=" mt-auto items-center p-4 bg-neutral text-neutral-content justify-center text-muted">
-	<p class="text-xs text-center">
-		Copyright © {year} - All right reserved
-	</p>
-</footer>
