+++ src/lib/components/ui/scrollable-table.svelte
