+++ src/routes/(calculators)/fire-calculator/_components/fire-chart.svelte
@@ -6,16 +6,18 @@
 		LineController,
 		LineElement,
 		PointElement,
-		type ChartDataset,
 		CategoryScale,
 		LinearScale,
 		Legend,
 		Tooltip,
-		Filler
+		Filler,
+		BarController,
+		BarElement,
+		type ChartDataset
 	} from 'chart.js';
 	import { formatAsCurrency } from '$lib/utils/formatters';
 	import { COLOURS } from '$lib/constants/colours';
-	import { TOOLTIP } from '$lib/constants/chartConfig';
+	import { TOOLTIP } from '$constants/chart-config';
 	import { getCalculatorState } from '../calculator.svelte';
 
 	const calculator = getCalculatorState();
@@ -30,16 +32,21 @@
 		LinearScale,
 		Legend,
 		Tooltip,
-		Filler
+		Filler,
+		BarController,
+		BarElement
 	);
 
-	function createDatasets(yearlyData: typeof calculator.outcome.yearlyData) {
-		// Create datasets array starting with superannuation
-		const datasets = [
+	type MixedDataset =
+		| (ChartDataset<'line', number[]> & { type: 'line' })
+		| (ChartDataset<'bar', number[]> & { type: 'bar' });
+
+	function createDatasets(calculationData: typeof calculator.calculationData) {
+		const baseDatasets: MixedDataset[] = [
 			{
-				type: 'line' as const,
-				label: 'Super Balance',
-				data: yearlyData.map((d) => (d.superannuationBalance > 0 ? d.superannuationBalance : null)),
+				type: 'line',
+				label: 'Investment Value',
+				data: calculationData.map((d) => d.investmentValue),
 				borderColor: COLOURS[0],
 				backgroundColor: `${COLOURS[0]}33`,
 				fill: true,
@@ -47,48 +54,84 @@
 				borderWidth: 2,
 				pointRadius: 0,
 				pointHoverRadius: 4,
+				order: 3
+			},
+			{
+				type: 'line',
+				label: 'FIRE Target',
+				data: calculationData.map((d) => d.fireTarget),
+				borderColor: COLOURS[1],
+				borderDash: [5, 5],
+				borderWidth: 2,
+				pointRadius: 0,
+				fill: false,
 				order: 1
 			}
 		];
 
-		// If we have at least one data point with investment breakdown data
-		if (yearlyData.length > 0 && yearlyData[0].investmentsBreakdown) {
-			// Instead of grouping all investments, create individual line for each investment
-			const investmentNames = yearlyData[0].investmentsBreakdown.map((inv) => inv.name);
+		const withdrawalDatasets: MixedDataset[] = [];
 
-			// For each investment, create a line
-			investmentNames.forEach((name, index) => {
-				// Calculate color index to avoid color collisions
-				const colorIndex = (index + 1) % COLOURS.length;
+		if (calculator.secondaryIncome > 0) {
+			// Add net withdrawals (withdrawals minus secondary income)
+			withdrawalDatasets.push({
+				type: 'bar',
+				label: 'Withdrawals',
+				data: calculationData.map((d) => d.withdrawal - d.secondaryIncome),
+				borderColor: COLOURS[2],
+				backgroundColor: `${COLOURS[2]}33`,
 
-				datasets.push({
-					type: 'line' as const,
-					label: name,
-					data: yearlyData.map((d) => {
-						const investment = d.investmentsBreakdown.find((inv) => inv.name === name);
-						return investment && investment.value > 0 ? investment.value : null;
-					}),
-					borderColor: COLOURS[colorIndex],
-					backgroundColor: `${COLOURS[colorIndex]}33`,
-					fill: true,
-					tension: 0.4,
-					borderWidth: 2,
-					pointRadius: 0,
-					pointHoverRadius: 4,
-					order: index + 2
-				});
+				borderWidth: 1,
+				borderRadius: 2,
+
+				order: 2,
+				barPercentage: 0.8,
+				stack: 'expenses'
+			});
+
+			// Add secondary income bar
+			withdrawalDatasets.push({
+				type: 'bar',
+				label: 'Secondary Income',
+				data: calculationData.map((d) => d.secondaryIncome),
+				borderColor: COLOURS[3] || colors.green[500],
+				backgroundColor: `${COLOURS[3]}33`,
+
+				borderWidth: 1,
+				borderRadius: 2,
+
+				order: 2,
+				barPercentage: 0.8,
+				stack: 'expenses'
+			});
+		} else {
+			// If no secondary income, just show total withdrawals
+			withdrawalDatasets.push({
+				type: 'bar',
+				label: 'Withdrawals',
+				data: calculationData.map((d) => d.withdrawal),
+				borderColor: COLOURS[2],
+				backgroundColor: `${COLOURS[2]}33`,
+
+				borderWidth: 1,
+				borderRadius: 2,
+
+				order: 2,
+				barPercentage: 0.8,
+				stack: 'expenses'
 			});
 		}
 
-		return datasets;
+		const dataset: MixedDataset[] = [...baseDatasets, ...withdrawalDatasets];
+
+		return dataset;
 	}
 
 	onMount(() => {
 		chart = new Chart(chartId!, {
 			type: 'line',
 			data: {
-				labels: calculator.outcome.yearlyData.map((d) => d.age),
-				datasets: createDatasets(calculator.outcome.yearlyData)
+				labels: calculator.calculationData.map((d) => d.age),
+				datasets: createDatasets(calculator.calculationData)
 			},
 			options: {
 				maintainAspectRatio: false,
@@ -122,8 +165,8 @@
 							color: colors.slate[200]
 						},
 						title: {
-							display: true,
-							text: 'Asset Balance',
+							display: false,
+							text: 'Balance',
 							font: { size: 16, family: 'sans-serif' },
 							color: colors.slate[200]
 						}
@@ -140,7 +183,7 @@
 							title: (tooltip) => `Age ${tooltip[0].label}`,
 							label: (context) => {
 								const label = context.dataset.label || '';
-								const value = context.parsed.y || 0;
+								const value = Math.abs(context.parsed.y || 0);
 								return `${label}: ${formatAsCurrency(value)}`;
 							}
 						}
@@ -167,18 +210,18 @@
 	});
 
 	$effect(() => {
-		if (chart && calculator.outcome.yearlyData) {
-			// Update datasets with new data
-			chart.data.datasets = createDatasets(calculator.outcome.yearlyData);
+		if (chart) {
+			// Always start with base dataset
+			chart.data.datasets = createDatasets(calculator.calculationData);
 
 			// Update labels
-			chart.data.labels = calculator.outcome.yearlyData.map((d) => d.age);
+			chart.data.labels = calculator.calculationData.map((d) => d.age);
 
 			chart.update();
 		}
 	});
 </script>
 
-<div class="min-h-[300px] lg:min-h-[350px] relative">
+<div class="min-h-[400px] lg:min-h-[500px] relative">
 	<canvas class="w-full absolute min-h-full p-1" bind:this={chartId}></canvas>
 </div>
