+++ src/routes/(calculators)/fire-calculator/_components/fire-inputs.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
 	import { getCalculatorState } from '../calculator.svelte';
-	import PercentageInput from '$lib/components/inputs/PercentageInput.svelte';
-	import CurrencyInput from '$lib/components/inputs/CurrencyInput.svelte';
+	import PercentageInput from '$lib/components/inputs/percentage-input.svelte';
+	import CurrencyInput from '$lib/components/inputs/currency-input.svelte';
 	import Label from '$ui/label/label.svelte';
 	import Input from '$ui/input/input.svelte';
 	import { Button } from '$lib/components/ui/button';
 	import Separator from '$lib/components/Separator.svelte';
-	import Explainer from '$lib/components/Explainer.svelte';
+	import Explainer from '$lib/components/explainer.svelte';
 
 	let calculator = getCalculatorState();
 	let baristaFi = $state(false);
