+++ src/routes/(calculators)/fire-calculator/_components/fire-table.svelte
