+++ src/routes/(calculators)/mortgage-calculator/_components/mortgage-chart.svelte
@@ -13,23 +13,12 @@
 		Filler
 	} from 'chart.js';
 	import { formatAsCurrency } from '$lib/utils/formatters';
-	import type { AnnualData } from '../calculator.svelte';
 	import { COLOURS } from '$lib/constants/colours';
-	import { TOOLTIP } from '$lib/constants/chartConfig';
+	import { TOOLTIP } from '$constants/chart-config';
+	import { getCalculatorState } from '../calculator.svelte';
 
-	let {
-		baseData = [],
-		comparisonData = [],
-		savingsGoal,
-		isComparing
-	}: {
-		baseData: AnnualData[];
-		comparisonData: AnnualData[];
-		savingsGoal: number;
-		isComparing: boolean;
-	} = $props();
+	const calculator = getCalculatorState();
 
-	let years = $derived(baseData.map((item) => item.year));
 	let chartId: HTMLCanvasElement | undefined = $state();
 	let chart: Chart | undefined = $state();
 
@@ -44,11 +33,35 @@
 		Filler
 	);
 
-	// Helper function to create dataset
-	function createDataset(data: AnnualData[], color: string, label: string, options = {}) {
+	let years = $derived(
+		Array.from(
+			{ length: Math.ceil(calculator.totalPeriods / calculator.periodsPerYear) },
+			(_, i) => i + 1
+		)
+	);
+
+	function getSafePeriodData(periodData: any[], index: number) {
+		return periodData && periodData[index]
+			? periodData[index]
+			: { balance: 0, totalInterestPaid: 0 };
+	}
+
+	function createDataset(
+		valueKey: 'balance' | 'totalInterestPaid' | 'propertyValue',
+		color: string,
+		label: string,
+		options = {}
+	) {
+		const periodData = calculator.projection.breakdown || [];
+
+		const data = Array.from({ length: years.length }, (_, i) => {
+			const periodIndex = (i + 1) * calculator.periodsPerYear - 1;
+			return getSafePeriodData(periodData, periodIndex)[valueKey];
+		});
+
 		return {
 			label,
-			data: Array.from({ length: years.length }, (_, i) => data[i].endingValue),
+			data,
 			borderColor: color,
 			backgroundColor: color + '40',
 			fill: true,
@@ -60,30 +73,17 @@
 		};
 	}
 
-	function createGoalDataset(value: number, color: string) {
-		return {
-			label: 'Goal',
-			data: Array.from({ length: years.length }, () => value),
-			borderColor: color,
-			backgroundColor: 'transparent',
-			fill: false,
-			borderWidth: 2,
-			borderDash: [5, 5],
-			pointRadius: 0,
-			pointHoverRadius: 4,
-			pointBackgroundColor: color
-		};
-	}
-
 	onMount(() => {
 		chart = new Chart(chartId!, {
 			type: 'line',
 			data: {
 				labels: years,
 				datasets: [
-					createDataset(baseData, COLOURS[0], 'Value'),
-					...(isComparing ? [createDataset(comparisonData, COLOURS[1], 'Comparison')] : []),
-					...(savingsGoal ? [createGoalDataset(savingsGoal, COLOURS[2])] : [])
+					createDataset('balance', COLOURS[0], 'Loan Balance'),
+					createDataset('totalInterestPaid', COLOURS[1], 'Total Interest'),
+					...(calculator.propertyValue > 0
+						? [createDataset('propertyValue', COLOURS[2], 'Property Value')]
+						: [])
 				]
 			},
 			options: {
@@ -108,6 +108,7 @@
 						display: true,
 						stacked: false,
 						position: 'left',
+						min: 0,
 						grid: {
 							display: true,
 							color: colors.slate[600]
@@ -120,7 +121,7 @@
 						},
 						title: {
 							display: false,
-							text: 'Balance',
+							text: 'Amount',
 							font: { size: 16, family: 'sans-serif' },
 							color: colors.slate[200]
 						}
@@ -164,23 +165,15 @@
 	});
 
 	$effect(() => {
-		if (chart) {
-			// Always start with base dataset
-			chart.data.datasets = [createDataset(baseData, COLOURS[0], 'Value')];
-
-			// Add comparison if active
-			if (isComparing) {
-				chart.data.datasets.push(createDataset(comparisonData, COLOURS[1], 'Comparison'));
-			}
-
-			// Add goal if exists
-			if (savingsGoal) {
-				chart.data.datasets.push(createGoalDataset(savingsGoal, COLOURS[2]));
-			}
-
-			// Update labels
+		if (chart && chart.data) {
+			chart.data.datasets = [
+				createDataset('balance', COLOURS[0], 'Loan Balance'),
+				createDataset('totalInterestPaid', COLOURS[1], 'Total Interest'),
+				...(calculator.propertyValue > 0
+					? [createDataset('propertyValue', COLOURS[2], 'Property Value')]
+					: [])
+			];
 			chart.data.labels = years;
-
 			chart.update();
 		}
 	});
