+++ src/routes/(calculators)/portfolio-builder/+page.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
-	import CurrencyInput from '$lib/components/inputs/CurrencyInput.svelte';
+	import CurrencyInput from '$lib/components/inputs/currency-input.svelte';
 	import { formatAsCurrency, formatAsPercentage } from '$utils/formatters';
-	import AddInvestment from './_components/AddInvestment.svelte';
-	import InvestmentItem from './_components/InvestmentItem.svelte';
+	import AddInvestment from './_components/add-investment.svelte';
+	import InvestmentItem from './_components/investment-item.svelte';
 	import { setPortfolioState } from './calculator.svelte';
-	import DownloadButton from '$lib/components/DownloadButton.svelte';
+	import DownloadButton from '$lib/components/download-button.svelte';
 	import * as Tabs from '$lib/components/ui/tabs';
-	import AssetAllocationChart from './_components/AssetAllocationChart.svelte';
-	import AssetAllocationTable from './_components/AssetAllocationTable.svelte';
+	import AssetAllocationChart from './_components/allocation-chart.svelte';
+	import AssetAllocationTable from './_components/allocation-table.svelte';
 
 	const calc = setPortfolioState();
