+++ src/routes/(calculators)/portfolio-builder/_components/add-investment.svelte
