+++ src/routes/(calculators)/portfolio-builder/_components/allocation-table.svelte
