+++ src/routes/(calculators)/retirement-projection/_components/outcome-card.svelte
