+++ src/routes/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import CalculatorCard from './_components/CalculatorCard.svelte';
+	import CalculatorCard from './_components/calculator-card.svelte';
 	import { calculators } from '$lib/constants/calculators';
 </script>
