+++ src/routes/_components/calculator-card.svelte
